Convert integers to unsigned integers where negative values don't make sense

+++ src/av_resources.cpp
@@ -53,7 +53,9 @@ WrappedAvPacket::~WrappedAvPacket()
 WrappedAvPacket::WrappedAvPacket(AVPacket *packet) : packet(packet) {}
 
 Expected<std::shared_ptr<WrappedAvFrame>, Error> WrappedAvFrame::create(
-    const enum AVPixelFormat pixel_format, const int width, const int height
+    const enum AVPixelFormat pixel_format,
+    const unsigned int width,
+    const unsigned int height
 )
 {
     AVFrame *frame = av_frame_alloc();
@@ -149,9 +151,9 @@ Expected<std::shared_ptr<WrappedAvCodecContext>, Error>
     WrappedAvCodecContext::create(
         const enum AVCodecID codec_id,
         const int64_t bit_rate,
-        const int width,
-        const int height,
-        const int frame_rate,
+        const unsigned int width,
+        const unsigned int height,
+        const unsigned int frame_rate,
         const enum AVPixelFormat source_pixel_format,
         const int additional_flags,
         const WrappedAvDictionary &parameters
@@ -188,7 +190,7 @@ Expected<std::shared_ptr<WrappedAvCodecContext>, Error>
     // of which frame timestamps are represented. For fixed-fps content,
     // timebase should be 1/framerate and timestamp increments should be
     // identical to 1.
-    codec_context->time_base = (AVRational){1, frame_rate};
+    codec_context->time_base = (AVRational){1, static_cast<int>(frame_rate)};
 
     // Emit one intra frame every twelve frames at most.
     codec_context->gop_size = 12;
@@ -274,9 +276,9 @@ Expected<std::shared_ptr<WrappedOutputVideoAvFormatContext>, Error>
         const std::string &filename,
         const std::optional<std::string> &format_name,
         const int64_t bit_rate,
-        const int width,
-        const int height,
-        const int frame_rate,
+        const unsigned int width,
+        const unsigned int height,
+        const unsigned int frame_rate,
         const enum AVPixelFormat source_pixel_format,
         const WrappedAvDictionary &parameters,
         const bool intermediate_yuv420p_conversion
